neubot/raw_srvr.py: save timestamp on the server
We save ticks in server's goodput, which is OK as long as the
server is Linux.  I contend that we only have Linux servers,
still better to write good code in any case.  So add a timestamp
to the results, which will work regardless the platform we use
on the server side.