web100: workaround apparently nonstandard IPv6 address repr
AFAIK IPv6 addresses should not containt triple colons: the maximum
is one occurrence of a double colon.  However, it turns out that
Web100 puts IPv6 addresses with triple colons in /web100/*/spec files.
Be it a problem with Web100 or my limited knowledge, better to add
a workaround, such that Neubot can read Web100 variables when tests
use the IPv6 protocol.