background_api: honor /etc/neubot/api
Such that someone can specify alternate address and port to
listen to.  This is just initial support for this feature.
More work is needed to migrate existing modules to that and
decommission agent.api.{address,port} settings.

I'd rather use a configuration file instead of a setting
in the database.  Indeed, the former does not need you
to do `black magic` to access Neubot's web UI to change
the default address and port.

Requested on Neubot's Facebook Page by Eric Speckman.