http_clnt: more pythonic event dispatch
This diff backports more pythonic event dispatch from the pu
branch (i.e., 0.5.0.x).  To be committed to master (i.e.,
0.4.16.x).

While there, it ensures that the HTTP reader raises an
exception when there are too many incoming headers.

Also, it performs some cosmetic changes needed to harmonize
the style of this file with http_srvr.py (which exists in
the pu branch only).

Finally, it propagate changes to files using http_clnt.py.

It does not update the regression test suite to test the
maximum-headers feature.  This is to be done in a separate
diff.