regress: add neubot/http_clnt.py: unit test for http_clnt.py
To be committed into the stable branch first (i.e., 0.4.15.x),
then ported to the master branch (i.e., 0.4.16.x).  The file
is not complete yet, because http code is changed significantly
in master.  So, I won't bother myself with writing tests for
stable and then adapt them to master.  Rather, I will add them
to master when I forward port this patch to it.