M-Lab: adapt to new deployment style
Now Measurement Lab provides some support for deploying an experiment
to slivers. This commit adapts existing code to their conventions.

Part of the code has been moved into a dedicated repository, which
is currently hosted at::

    https://github.com/bassosimone/neubot_mlab

To better understand this commit, it is particularly helpful to
consider the documentation of `neubot_mlab`::

    https://github.com/bassosimone/neubot_mlab/blob/master/documentation.png

The most relevant change is that we don't need anymore to start
syslogd and rsync, because this is now done automatically.

While writing this diff, I also added some extra checks; e.g., that
the scripts are invoked by root.