DASH test prototype - v1
This is the first version of the diff that implements
the new MPEG DASH test.

The logic of the test is still missing.

However, the infrastructure around the test is now
ready to be pushed to Github (still beta, eh).

This new test is as self contained as possible; the
idea of self-contained tests should be the first
step towards implementing real plugin network tests.

More to come.