neubot/__init__.py: update copyright years and wording
Specify more correctly that there is a dual copyright:

- the individual contributor (me in this case) owns the code; and

- the Nexa Center owns the code as well.

While there also make sure we write Nexa rather than NEXA.

I already did this in the past for individual files, but it is
handy to perform this change to this file, because I create
new files by copying this file. Therefore I don't have to edit
at hand the copyright statement each time.