Add the new dash test
The new test is not inside the Neubot directory, because this one
is the first test that is implemented as a plugin.

The new test is implemented as a plugin thanks to the new
functionality that I added to utils_modules.py.

Very few bits of Neubot will have the 'dash' name hardcoded.

This diff is joint work with Antonio Servetti (thanks!).