client_smpl.py: less aggressive bisect
At home I noticed that I download at 6.7 Mbit/s and the test
continues to request 7 Mbit/s of rate.

This is wrong, because instead the test should use 6 Mbit/s
as a rate when the speed is 6.7 Mbit/s.

This diff fixes the problem, by picking the element which is
at the left of the bisection point.

Tested from the Nexa Center with many close and far servers
to make sure the new code works both in normal conditions
and when we go past the boundaries of the vector.

While there tweak the code that prints the result of the
test to put the rate first, then the speed, because it
is simpler for a human (like me) to read the new output
(the rate is decided first, then comes the actual download
speed).