By default use the poller functionality of libneubot
This commit modifies Neubot to use by default the poller
functionality of libneubot, which is Neubot support library
written in C and based on libevent.

Ideally, the C library should perform the networking and should
sporadically call Python to perform more high level tasks;
however, this is not possible unless we replace the event loop
of Neubot as well, which is exactly what this commit does.