neubot.1.rst: better dash rate-adaptation descr
What matters is not the two seconds threshold, which is an
implementation detail, but that we request the next segment
using the rate representation that is close to the speed
with which we downloaded the previous chunk.

(There is also the case in which we overcome the two seconds
threshold, which is not documented as of now. However, the
"in the common case" is there exactly to hint that what we say
is not exactly the full story. I will try to come up with an
even better description of dashtest in the future.)