{neubot_asyncio => net}/connect.py
The connect.py module should use neubot_asyncio but should not
be part of the neubot_asyncio module.

With this commit I'm reversing the trend of putting all the files
in neubot/ directory.

To help people understanding how Neubot works, I'll commit to be
more tidy than I used to be.