Second implementation of plugins, early draft
This commit adds the second implementation of plugins, which
ameliorates the first implementation, released along with the
current stable version of Neubot.

It is an early draft, but good enough that I can continue to
work on plugins on top of this branch.

The general idea is to come out with a more flexible mechanism
that also allows to execute external programs.

To show that executing external programs is possible, I've
also added a plugin for running traceroute.