_eventloop.py: defer SystemExit at end of loop
I just noticed that the real asyncio does not break immediately
out of the loop on SystemExit.

This diff modifies my asyncio emulator to do the same.

While there, do not raise when the loop is closed and its run
forever method is called twice, because the real asyncio is not
raising an exception in this case (in Python 3.4).