Remove notifier_browser.py
In the quest to put under control again the Neubot tree, let's get
rid of "the most annoying notification strategy ever" (cit.)

This notification strategy was annoying in 2012, I can't imagine
how much it would drive people crazy in 2017.

Per the agreement with Nexa, I am creating the "final" and frozen
release of Neubot, and we will not need this notifier in there.